Tweaks for BPicture compliance
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@5082 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/private/servers/app/ServerFont.h
@@ -41,14 +41,14 @@ public:
 	ServerFont(const ServerFont &font);
 	~ServerFont(void);
 	font_direction Direction(void) { return fdirection; }
-	uint8 Encoding(void) { return fencoding; }
+	uint32 Encoding(void) { return fencoding; }
 	edge_info Edges(void) { return fedges; }
 	uint32 Flags(void) { return fflags; }
-	uint8 Spacing(void) { return fspacing; }
+	uint32 Spacing(void) { return fspacing; }
 	float Shear(void) { return fshear; }
 	float Rotation(void) { return frotation; }
 	float Size(void) { return fsize; }
-	uint16 Face(void) { return fface; }
+	uint32 Face(void) { return fface; }
 	uint32 CountGlyphs(void);
 	int32 CountTuned(void);
 	font_file_format FileFormat(void);
@@ -56,13 +56,13 @@ public:
 
 	void SetDirection(const font_direction &dir) { fdirection=dir; }
 	void SetEdges(const edge_info &info) { fedges=info; }
-	void SetEncoding(uint8 encoding) { fencoding=encoding; }
+	void SetEncoding(uint32 encoding) { fencoding=encoding; }
 	void SetFlags(const uint32 &value) { fflags=value; }
-	void SetSpacing(const uint8 &value) { fspacing=value; }
+	void SetSpacing(const uint32 &value) { fspacing=value; }
 	void SetShear(const float &value) { fshear=value; }
 	void SetSize(const float &value) { fsize=value; }
 	void SetRotation(const float &value) { frotation=value; }
-	void SetFace(const uint16 &value) { fface=value; }
+	void SetFace(const uint32 &value) { fface=value; }
 	
 	BRect BoundingBox(void);
 	void Height(font_height *fh);
@@ -76,11 +76,11 @@ protected:
 	float fsize, frotation, fshear;
 	BRect fbounds;
 	uint32 fflags;
-	uint8 fspacing;
-	uint16 fface;
+	uint32 fspacing;
+	uint32 fface;
 	font_direction fdirection;
 	uint8 ftruncate;
-	uint8 fencoding;
+	uint32 fencoding;
 };
 
 #endif
